gateway/requirements.txt
paho-mqtt>=1.5.0
pyserial>=3.4
